+++ android/src/com/unciv/app/AndroidFont.kt
@@ -12,10 +12,10 @@ import android.os.Build
 import androidx.annotation.RequiresApi
 import com.badlogic.gdx.Gdx
 import com.badlogic.gdx.graphics.Pixmap
-import com.unciv.ui.components.FontFamilyData
-import com.unciv.ui.components.FontImplementation
-import com.unciv.ui.components.FontMetricsCommon
-import com.unciv.ui.components.Fonts
+import com.unciv.ui.components.fonts.FontFamilyData
+import com.unciv.ui.components.fonts.FontImplementation
+import com.unciv.ui.components.fonts.FontMetricsCommon
+import com.unciv.ui.components.fonts.Fonts
 import com.unciv.utils.Log
 import java.util.Locale
 import kotlin.math.abs
