+++ core/src/com/unciv/models/ruleset/nation/Difficulty.kt
@@ -5,7 +5,7 @@ import com.unciv.models.ruleset.unique.Unique
 import com.unciv.models.stats.INamed
 import com.unciv.ui.screens.civilopediascreen.FormattedLine
 import com.unciv.ui.screens.civilopediascreen.ICivilopediaText
-import com.unciv.ui.components.Fonts
+import com.unciv.ui.components.fonts.Fonts
 
 class Difficulty: INamed, ICivilopediaText {
     override lateinit var name: String
