+++ core/src/com/unciv/models/ruleset/tech/Era.kt
@@ -8,7 +8,7 @@ import com.unciv.models.ruleset.RulesetObject
 import com.unciv.models.ruleset.unique.StateForConditionals
 import com.unciv.models.ruleset.unique.UniqueTarget
 import com.unciv.models.ruleset.unique.UniqueType
-import com.unciv.ui.components.Fonts
+import com.unciv.ui.components.fonts.Fonts
 import com.unciv.ui.components.extensions.colorFromRGB
 import com.unciv.ui.screens.civilopediascreen.FormattedLine
