+++ core/src/com/unciv/models/translations/Translations.kt
@@ -7,7 +7,7 @@ import com.unciv.models.ruleset.RulesetCache
 import com.unciv.models.ruleset.unique.Unique
 import com.unciv.models.stats.Stat
 import com.unciv.models.stats.Stats
-import com.unciv.ui.components.Fonts
+import com.unciv.ui.components.fonts.Fonts
 import com.unciv.utils.Log
 import com.unciv.utils.debug
 import java.util.Locale
