+++ core/src/com/unciv/ui/screens/overviewscreen/UnitOverviewTab.kt
@@ -17,7 +17,7 @@ import com.unciv.models.UnitActionType
 import com.unciv.models.UpgradeUnitAction
 import com.unciv.models.ruleset.unit.BaseUnit
 import com.unciv.ui.components.widgets.ExpanderTab
-import com.unciv.ui.components.Fonts
+import com.unciv.ui.components.fonts.Fonts
 import com.unciv.ui.components.widgets.TabbedPager
 import com.unciv.ui.components.widgets.UnitGroup
 import com.unciv.ui.components.extensions.addSeparator
